Refactor: 커스텀 훅 네이밍 수정
